roller: move polling to it's own file, change polling behavior to be less request heavy
